Create how_many.cpp
Given the array nums, for each nums[i] find out how many numbers in the array are smaller than it. 
That is, for each nums[i] you have to count the number of valid j's such that j != i and nums[j] < nums[i].

Return the answer in an array.

unfortunately, this one didn't work. the one that worked previously, although using a different method, can be found in the following link: